pyproject.toml
[tool.poetry]
name = "robusta_actions"
version = "0.0.1"
description = ""
authors = ["Rivtower Technologies <contact@rivtower.com>"]

[tool.poetry.dependencies]
kubernetes = "^12.0.1"
openai = "^0.26.0"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.10.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"